+++ src/baobab_mtg_products/exceptions/product/__init__.py
@@ -1,5 +1,8 @@
 """Exceptions liées au modèle produit."""
 
+from baobab_mtg_products.exceptions.product.duplicate_internal_barcode_error import (
+    DuplicateInternalBarcodeError,
+)
 from baobab_mtg_products.exceptions.product.invalid_commercial_barcode_error import (
     InvalidCommercialBarcodeError,
 )
@@ -15,6 +18,9 @@ from baobab_mtg_products.exceptions.product.invalid_product_instance_error impor
 from baobab_mtg_products.exceptions.product.invalid_product_reference_error import (
     InvalidProductReferenceError,
 )
+from baobab_mtg_products.exceptions.product.invalid_production_code_error import (
+    InvalidProductionCodeError,
+)
 from baobab_mtg_products.exceptions.product.invalid_product_reference_id_error import (
     InvalidProductReferenceIdError,
 )
@@ -26,11 +32,13 @@ from baobab_mtg_products.exceptions.product.invalid_set_code_error import (
 )
 
 __all__ = [
+    "DuplicateInternalBarcodeError",
     "InvalidCommercialBarcodeError",
     "InvalidInternalBarcodeError",
     "InvalidProductIdentifierError",
     "InvalidProductInstanceError",
     "InvalidProductReferenceError",
+    "InvalidProductionCodeError",
     "InvalidProductReferenceIdError",
     "InvalidSerialNumberError",
     "InvalidSetCodeError",